* [Platform/Search] Fixed an issue when attempting to run cron.maint to purge deleted records before the full-text search indexes have been built.

+++ libs/devblocks/api/services/search.php
@@ -1231,6 +1231,8 @@ class DevblocksSearchEngineMysqlFulltext extends Extension_DevblocksSearchEngine
 	
 	public function delete(Extension_DevblocksSearchSchema $schema, $ids) {
 		$db = DevblocksPlatform::getDatabaseService();
+		$tables = DevblocksPlatform::getDatabaseTables();
+		
 		$ns = $schema->getNamespace();
 		
 		if(!is_array($ids))
@@ -1239,8 +1241,13 @@ class DevblocksSearchEngineMysqlFulltext extends Extension_DevblocksSearchEngine
 		if(empty($ns) || empty($ids))
 			return;
 			
+		$namespace = $this->escapeNamespace($ns);
+		
+		if(!isset($tables['fulltext_'.$namespace]))
+			return true;
+		
 		$result = $db->ExecuteMaster(sprintf("DELETE FROM fulltext_%s WHERE id IN (%s) ",
-			$this->escapeNamespace($ns),
+			$namespace,
 			implode(',', $ids)
 		));
